handle apis from next apis
