+++ src/app/interfaces/home.ts
@@ -1,0 +1,4 @@
+export type IQuote = {
+	text: string;
+	author: string;
+};
