--- src/app/types/home.ts
@@ -1,4 +1,0 @@
-export type QuoteType = {
-  text: string;
-  author: string;
-};
