+++ src/app/utils/index.ts
@@ -1,0 +1,4 @@
+export const getRandomIndexOfTheArray = (arr: any[]): number => {
+	const randomIndex: number = Math.floor(Math.random() * arr.length);
+	return randomIndex;
+};
