+++ src/app/utils/quote.ts
@@ -1,10 +1,6 @@
-import { QuoteType } from '../types/home';
+import { IQuote } from '../interfaces/home';
+import { getRandomIndexOfTheArray } from './index';
 
-export const getRandomIndexOfTheArray = (arr: any[]): number => {
-	const randomIndex: number = Math.floor(Math.random() * arr.length);
-	return randomIndex;
-};
-
-export const getRandomQuote = (arr: any[]): QuoteType => {
+export const getRandomQuote = (arr: any[]): IQuote => {
 	return arr[getRandomIndexOfTheArray(arr)];
 };
